intern-project/tasnim_siddique/images/requirements.txt
# Python deps for DAGs (Airflow already in image)
apache-airflow-providers-amazon==8.25.0
boto3>=1.28.0
s3fs>=2024.3.1
pandas>=2.2.0
pyarrow>=15.0.0
minio>=7.2.5
